+++ backend/youtube_tools/handle_cache.json
@@ -3,5 +3,6 @@
   "aquariusgirl_0172": "UCbDWjj0pzq2sZY51fJhxqVg",
   "kuro777": "UCk4ZazRDVOZfr6QYkE62kZw",
   "iamakich": "UCz07NnURsnxosjfM3mKYTGQ",
-  "Mollis_SamanV": "UC-efxH6WiBaVxsHq5RGcZkA"
+  "Mollis_SamanV": "UC-efxH6WiBaVxsHq5RGcZkA",
+  "bwslime": "UCroAcEtu4HwSIeKpULQwblw"
 }
