+++ backend/youtube_tools/youtube_channel_import.py
@@ -19,56 +19,33 @@ python youtube_channel_import.py
 from __future__ import annotations
 
 import argparse
-import json
-import logging
-import os
-import re
 import sys
-from pathlib import Path
 from typing import Dict, List
-
+from pathlib import Path
 from google.cloud import firestore
-from google.oauth2 import service_account
-from googleapiclient.discovery import build
-from googleapiclient.errors import HttpError
 from dotenv import load_dotenv
 
+from core.handle_utils import parse_and_resolve_channel_ids
+from core.youtube_api import (
+    build_youtube_service,
+    fetch_channels_info
+)
+from core.constants import (
+    get_api_key, get_firebase_key_path,
+    LOG_FILE,
+    FIRESTORE_INFO_PATH, FIRESTORE_INDEX_COLLECTION, SPECIAL_CHANNEL_ID,
+)
+from core.firestore_writer import init_firestore_client, needs_update_info
+from core.log_setup import logger
+
 # ---------------------------------------------------------------------------#
-# 📂 環境變數與路徑
+# 📂 載入環境變數
 # ---------------------------------------------------------------------------#
 if not load_dotenv("../.env.local"):
     load_dotenv(".env.local")
 
-BASE_DIR = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
-
-YOUTUBE_API_KEY = os.getenv("API_KEY")
-FIREBASE_KEY_PATH = os.getenv("FIREBASE_KEY_PATH", os.path.join(BASE_DIR, "firebase-key.json"))
-
-# ---------------------------------------------------------------------------#
-# 設定常數
-# ---------------------------------------------------------------------------#
-HANDLES_FILE = Path("channel_list_handle.txt")
-CACHE_FILE = Path("handle_cache.json")
-LOG_FILE = Path("youtube_channel_import.log")
-FIRESTORE_INFO_PATH = "channel_data/{channel_id}/channel_info/info"
-FIRESTORE_INDEX_COLLECTION = "channel_index"
-SPECIAL_CHANNEL_ID = "UCLxa0YOtqi8IR5r2dSLXPng"
-
-YT_CHANNELS_ENDPOINT_PARTS = "snippet"
-YT_CHANNELS_MAX_BATCH = 50
-
-# ---------------------------------------------------------------------------#
-# 日誌設定
-# ---------------------------------------------------------------------------#
-logger = logging.getLogger("youtube_channel_import")
-logger.setLevel(logging.INFO)
-_fmt = logging.Formatter("%(asctime)s | %(levelname)s | %(message)s", "%Y-%m-%d %H:%M:%S")
-sh = logging.StreamHandler()
-sh.setFormatter(_fmt)
-logger.addHandler(sh)
-fh = logging.FileHandler(LOG_FILE, encoding="utf-8")
-fh.setFormatter(_fmt)
-logger.addHandler(fh)
+YOUTUBE_API_KEY = get_api_key()
+FIREBASE_KEY_PATH = get_firebase_key_path()
 
 # ---------------------------------------------------------------------------#
 # 參數
@@ -79,137 +56,19 @@ def parse_args() -> argparse.Namespace:
     return parser.parse_args()
 
 # ---------------------------------------------------------------------------#
-# YouTube API
-# ---------------------------------------------------------------------------#
-HANDLE_REGEX = re.compile(
-    r"(?:https?://)?(?:www\.)?youtube\.com/(?:(?:channel/)?(?P<id>UC[0-9A-Za-z_-]{22,})|@?(?P<handle>[A-Za-z0-9_.-]+))/?"
-)
-
-def build_youtube_service(api_key: str):
-    return build("youtube", "v3", developerKey=api_key, cache_discovery=False)
-
-def resolve_handle_to_id(api, handle: str) -> str | None:
-    """
-    將 @handle 轉成 UCXXXX channel ID。
-    1) 先用 channels().list(forHandle=...) 精準對應
-    2) 若 404 / 400，退回 search().list(q=handle) 取第一筆
-    """
-    cleaned = handle.lstrip("@")
-    try:
-        resp = api.channels().list(part="id", forHandle=cleaned).execute()
-        items = resp.get("items", [])
-        if items:
-            return items[0]["id"]
-    except HttpError as e:
-        if e.resp.status not in (400, 404):
-            logger.error("[forHandle 失敗] handle=%s，原因：%s", handle, e)
-    except Exception as e:
-        logger.error("[forHandle 例外] handle=%s，原因：%s", handle, e)
-
-    try:
-        resp = api.search().list(part="snippet", type="channel", q=handle, maxResults=1).execute()
-        items = resp.get("items", [])
-        if items:
-            return items[0]["snippet"]["channelId"]
-    except Exception as e:
-        logger.error("[search 失敗] handle=%s，原因：%s", handle, e)
-
-    return None
-
-def pick_best_thumbnail(thumbnails: Dict) -> str:
-    for key in ("maxres", "standard", "high", "medium", "default"):
-        if key in thumbnails and thumbnails[key].get("url"):
-            return thumbnails[key]["url"]
-    return ""
-
-def fetch_channels_info(api, ids: List[str]) -> Dict[str, Dict]:
-    info: Dict[str, Dict] = {}
-    for i in range(0, len(ids), YT_CHANNELS_MAX_BATCH):
-        batch = ids[i : i + YT_CHANNELS_MAX_BATCH]
-        try:
-            resp = api.channels().list(part=YT_CHANNELS_ENDPOINT_PARTS, id=",".join(batch)).execute()
-        except Exception as e:
-            logger.error("[抓取失敗] 批次 %s，原因：%s", batch, e)
-            continue
-        for item in resp.get("items", []):
-            cid = item["id"]
-            snippet = item.get("snippet", {})
-            info[cid] = {
-                "name": snippet.get("title", ""),
-                "thumbnail": pick_best_thumbnail(snippet.get("thumbnails", {})),
-            }
-    return info
-
-# ---------------------------------------------------------------------------#
-# Firestore
-# ---------------------------------------------------------------------------#
-def init_firestore_client() -> firestore.Client:
-    credentials = service_account.Credentials.from_service_account_file(FIREBASE_KEY_PATH)
-    return firestore.Client(credentials=credentials, project=credentials.project_id)
-
-def needs_update_info(existing: Dict | None, new: Dict) -> bool:
-    if not existing:
-        return True
-    return (
-        existing.get("name") != new["name"]
-        or existing.get("thumbnail") != new["thumbnail"]
-        or existing.get("url") != new["url"]
-    )
-
-# ---------------------------------------------------------------------------#
 # 主流程
 # ---------------------------------------------------------------------------#
 def main():
     if not YOUTUBE_API_KEY or not Path(FIREBASE_KEY_PATH).exists():
         logger.critical("❌ 環境變數缺少 YOUTUBE_API_KEY 或找不到 FIREBASE_KEY_PATH，請檢查設定")
         sys.exit(1)
-    if not HANDLES_FILE.exists():
-        logger.critical("❌ 找不到 %s，無法繼續", HANDLES_FILE)
-        sys.exit(1)
 
     args = parse_args()
     youtube = build_youtube_service(YOUTUBE_API_KEY)
     fs_client = init_firestore_client()
 
-    # 讀取快取
-    handle_cache: Dict[str, str] = {}
-    if CACHE_FILE.exists():
-        try:
-            handle_cache = json.loads(CACHE_FILE.read_text(encoding="utf-8"))
-        except json.JSONDecodeError:
-            logger.warning("⚠️ 快取檔案已損毀，重新初始化")
-
-    # 解析輸入清單
-    logger.info("[🔎開始解析] ------ ")
-    channel_ids: List[str] = []
-    for line in HANDLES_FILE.read_text(encoding="utf-8").splitlines():
-        line = line.strip()
-        if not line:
-            continue
-        m = HANDLE_REGEX.match(line)
-        if not m:
-            logger.warning("⚠️ 無法解析：%s", line)
-            continue
-        cid, handle = m.group("id"), m.group("handle")
-        if cid:
-            channel_ids.append(cid)
-        elif handle:
-            channel_ids.append(handle_cache.get(handle) or handle)
-
-    # 解析 @handle → channel ID
-    for h in [h for h in channel_ids if not h.startswith("UC")]:
-        cid = resolve_handle_to_id(youtube, h)
-        if cid:
-            handle_cache[h] = cid
-            channel_ids[channel_ids.index(h)] = cid
-            logger.info("[解析完成] %s → %s", h, cid)
-        else:
-            logger.error("[解析失敗] %s", h)
-
-    CACHE_FILE.write_text(json.dumps(handle_cache, ensure_ascii=False, indent=2), encoding="utf-8")
-    # 去重
-    channel_ids = list(dict.fromkeys([cid for cid in channel_ids if cid.startswith("UC")]))
-    logger.info("待處理頻道：%d", len(channel_ids))
+    # 處理 handle 與快取 → 回傳 UC ID 清單
+    channel_ids = parse_and_resolve_channel_ids(youtube)
 
     # 取得頻道資訊
     info_map = fetch_channels_info(youtube, channel_ids)
@@ -265,6 +124,7 @@ def main():
         success.append(cid)
         result_map[cid] = {"name": data["name"], "status": status}
 
+    # 匯出摘要
     logger.info("----- 匯入摘要 -----")
     logger.info("成功：%d", len(success))
     logger.info("失敗：%d", len(failed))
